Using EfficientNet in the pipeline
