+++ requirements.txt
@@ -1,4 +1,3 @@
-altgraph @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
 anyio==4.0.0
 appnope==0.1.3
 argon2-cffi==23.1.0
@@ -26,7 +25,6 @@ executing==1.2.0
 fastjsonschema==2.18.0
 filelock==3.12.2
 fqdn==1.5.1
-future @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/future-0.18.2-py3-none-any.whl
 idna==3.4
 importlib-metadata==6.8.0
 ipykernel==6.25.2
@@ -52,7 +50,6 @@ jupyterlab==4.0.6
 jupyterlab-pygments==0.2.2
 jupyterlab-widgets==3.0.9
 jupyterlab_server==2.25.0
-macholib @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
 MarkupSafe==2.1.3
 matplotlib-inline==0.1.6
 mistune==3.0.1
@@ -91,7 +88,6 @@ rfc3339-validator==0.1.4
 rfc3986-validator==0.1.1
 rpds-py==0.10.3
 Send2Trash==1.8.2
-six @ file:///AppleInternal/Library/BuildRoots/9dd5efe2-7fad-11ee-b588-aa530c46a9ea/Library/Caches/com.apple.xbs/Sources/python3/six-1.15.0-py2.py3-none-any.whl
 sniffio==1.3.0
 soupsieve==2.5
 sqlparse==0.4.4
@@ -113,4 +109,7 @@ websocket-client==1.6.3
 widgetsnbextension==4.0.9
 zipp==3.17.0
 tqdm==4.66.2
-scikit-learn==1.4.1
+scikit-learn==1.4.1.post1
+efficientnet_pytorch==0.7.1
+torch==2.2.1
+torchvision==0.17.1
